Recurring-task occurrence history, member email login, faster Worker reads
Track per-day task_occurrences so recurring tasks keep a completion/history audit and today's check still renders after the rolling row advances.

Add member login with temporary passwords and welcome/reset emails via Resend.

Fix Worker 503s: watermark + batched occurrence inserts, batched dashboard member queries, and a short-TTL weather cache.

Refresh client branding/logos and add matching drizzle migrations.

+++ client/src/modules/calendar/CalendarPage.jsx
@@ -98,9 +98,9 @@ export default function CalendarPage() {
 
   let rangeLabel;
   if (view === "day") {
-    rangeLabel = anchor.toLocaleDateString(undefined, { weekday: "long", month: "long", day: "numeric", year: "numeric" });
+    rangeLabel = anchor.toLocaleDateString(undefined, { weekday: "long", month: "short", day: "numeric", year: "numeric" });
   } else if (view === "month") {
-    rangeLabel = start.toLocaleDateString(undefined, { month: "long", year: "numeric" });
+    rangeLabel = start.toLocaleDateString(undefined, { month: "short", year: "numeric" });
   } else {
     rangeLabel = `${fmtDate(start).slice(5).replace("-", "/")} – ${fmtDate(addDays(start, 6)).slice(5).replace("-", "/")}`;
   }
